Requesting Signoff for Revision of the GUI subsystem
I have changed the system to take in the user interaction as a high-priority interrupt. I believe that you still wanted me to mention the 300 ms delay but if you feel this is not necessary anymore, since the user interaction will be one of the highest interrupt priorities, I will remove it from the constraints and leave the user interrupt priority.